feat(setup.sh): make setup script compatible with working workstation

+++ setup.sh
@@ -4,10 +4,10 @@ set -e
 
 # determine linux distro
 if [ -f /etc/os-release ]; then
-    . /etc/os-release
-    OS=$ID
+  . /etc/os-release
+  OS=$ID
 else
-    OS=$(uname -s)
+  OS=$(uname -s)
 fi
 
 BREW_PATH="/home/linuxbrew/.linuxbrew/bin/brew"
@@ -16,34 +16,41 @@ DOTFILES_DIR=$(cd -- "$(dirname -- "$0")" && pwd)
 echo "🚀 Start setup for: $OS"
 
 echo "📦 Install system dependencies..."
+
 case "$OS" in
-    ubuntu|debian)
-        sudo apt update && sudo apt install -y build-essential stow git libssl-dev
-        ;;
-    fedora)
-        sudo dnf group install development-tools && sudo dnf install -y stow git curl openssl-devel dbus-daemon dbus-x11
-        ;;
+ubuntu | debian)
+  sudo apt update && sudo apt install -y build-essential stow git libssl-dev
+  ;;
+fedora)
+  sudo dnf group install development-tools && sudo dnf install -y stow git curl openssl-devel dbus-daemon dbus-x11
+  ;;
 esac
 
-if [[ ! -d $HOME/Documents ]]; then
-    echo "Creating Documents dir..."
-    mkdir $HOME/Documents
+if [[ ! -d "$HOME/Documents" ]]; then
+  echo "Creating Documents dir..."
+  mkdir "$HOME/Documents"
 fi
 
 # install homebrew if not installed
 if [[ ! -f "$BREW_PATH" ]]; then
-    echo "🍺 Installing Homebrew..."
-    /bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh)"
-    
-    eval "$($BREW_PATH shellenv bash)"
-    brew analytics off
-    brew install gcc
+  echo "🍺 Installing Homebrew..."
+  /bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh)"
+
+  eval "$($BREW_PATH shellenv bash)"
+  brew analytics off
+  brew install gcc
 else
-    eval "$($BREW_PATH shellenv)"
+  eval "$($BREW_PATH shellenv)"
 fi
 
 if [ -f "$DOTFILES_DIR/brew/Brewfile" ]; then
-    brew bundle --file "$DOTFILES_DIR/brew/Brewfile"
+  brew trust alesbrelih/gitlab-ci-ls
+  brew trust hashicorp/tap
+  brew trust robusta-dev/homebrew-krr
+  brew trust siderolabs/tap
+  brew trust terraform-linters/tap
+
+  brew bundle --file "$DOTFILES_DIR/brew/Brewfile"
 fi
 
 if [[ ! -d $HOME/.tmux/plugins/tpm ]]; then
@@ -52,8 +59,8 @@ if [[ ! -d $HOME/.tmux/plugins/tpm ]]; then
 fi
 
 if [ -f "$HOME/.gitconfig" ] && ! grep -q ".dotfiles_gitconfig" "$HOME/.gitconfig"; then
-    echo "📝 add .gitconfig..."
-    git config --global include.path "../.dotfiles_gitconfig"
+  echo "📝 add .gitconfig..."
+  git config --global include.path "../.dotfiles_gitconfig"
 fi
 
 cd "$DOTFILES_DIR" | exit
@@ -63,19 +70,23 @@ echo "🔗 Createting Symlinks with GNU Stow..."
 [ -d "$HOME/.config/nvim" ] && [ ! -L "$HOME/.config/nvim" ] && rm -rf "$HOME/.config/nvim"
 
 for dir in */; do
-    dir=${dir%/}
-    if [[ "$dir" != "freecad" && "$dir" != "brew" && "$dir" != ".git" ]]; then
-        echo "   Stowing $dir"
-        stow "$dir"
+  dir=${dir%/}
+  if [[ "$dir" != "freecad" && "$dir" != "brew" && "$dir" != ".git" ]]; then
+    if [[ "$dir" == "zsh" && "$USER" == "developer" ]]; then
+      ln -s "$HOME/dotfiles/zsh/.zshrc" "$HOME/.zsh/dotfiles-zshrc.sh"
+    else
+      echo "   Stowing $dir"
+      stow "$dir"
     fi
+  fi
 done
 
 FREECAD_TARGET="$HOME/.var/app/org.freecad.FreeCAD/config/FreeCAD/user.cfg"
 FREECAD_SOURCE="$DOTFILES_DIR/freecad/.var/app/org.freecadweb.FreeCAD/config/FreeCAD/user.cfg"
 
-if [ -f "$FREECAD_SOURCE" ]; then
-    mkdir -p "$(dirname "$FREECAD_TARGET")"
-    ln -sf "$FREECAD_SOURCE" "$FREECAD_TARGET"
+if [[ -f "$FREECAD_SOURCE" && "$USER" != "developer" ]]; then
+  mkdir -p "$(dirname "$FREECAD_TARGET")"
+  ln -sf "$FREECAD_SOURCE" "$FREECAD_TARGET"
 fi
 
 if [[ "$GIT_BIN" == "/home/linuxbrew/.linuxbrew/bin/git" ]]; then
@@ -91,11 +102,11 @@ fi
 # set zsh as default shell
 ZSH_PATH=$(command -v zsh)
 if [ "$SHELL" != "$ZSH_PATH" ]; then
-    echo "🐚 change default shell to zsh..."
-    if ! grep -q "$ZSH_PATH" /etc/shells; then
-        sudo bash -c "echo $ZSH_PATH >> /etc/shells"
-    fi
-    chsh -s "$ZSH_PATH"
+  echo "🐚 change default shell to zsh..."
+  if ! grep -q "$ZSH_PATH" /etc/shells; then
+    sudo bash -c "echo $ZSH_PATH >> /etc/shells"
+  fi
+  chsh -s "$ZSH_PATH"
 fi
 
 echo "✅ Setup finished!"
